fix(entity): added death sound and fixed variation bug (#3283)
* fix(entity): added death sound and fixed variation bug

* fix(entity): tests used creeper as a generic entity, which it is not

* fix clippy issue

* change from downcast to mob equivalent

* fix other clippy issues; followed same pattern from damage_with_context

* handle nautilus custom integration

* rename function

* fix clippy issue

* add custom_sound trait, fix guardian, magma_cube, slime, copper_golem and nautilus sounds; also fixed sound overlap between hurt and death sounds

* fix clippy

* cargo fmt

* remove unused imports

+++ crates/pumpkin/src/entity/player.rs
@@ -2219,7 +2219,7 @@ impl Player {
         position: &Vector3<f64>,
         volume: f32,
         pitch: f32,
-        seed: f64,
+        seed: i64,
     ) {
         let packet = CSoundEffect::new(IdOr::Id(sound_id), category, position, volume, pitch, seed);
         self.try_send_client_packet(&packet);
@@ -2232,7 +2232,7 @@ impl Player {
         position: &Vector3<f64>,
         volume: f32,
         pitch: f32,
-        seed: f64,
+        seed: i64,
     ) {
         let packet = CSoundEffect::new(IdOr::Value(sound), category, position, volume, pitch, seed);
         self.try_send_client_packet(&packet);
@@ -2267,7 +2267,7 @@ impl Player {
             position,
             volume,
             pitch,
-            rand::random::<f64>(),
+            rand::random::<i64>(),
         );
     }
