+++ crates/pumpkin/src/plugin/loader/wasm/wasm_host/wit/v0_1/player.rs
@@ -1636,7 +1636,7 @@ impl pumpkin::plugin::player::HostPlayer for PluginHostState {
             &pos,
             volume,
             pitch,
-            rand::random::<f64>(),
+            rand::random::<i64>(),
         );
         Ok(())
     }
@@ -1661,7 +1661,7 @@ impl pumpkin::plugin::player::HostPlayer for PluginHostState {
             &pumpkin_util::math::vector3::Vector3::new(pos.0, pos.1, pos.2),
             volume,
             pitch,
-            rand::random::<f64>(),
+            rand::random::<i64>(),
         );
         Ok(())
     }
